An smart app used by State Traffic Departments for tracking Vehicles
The Enhanced Check-Post Clearance and Fast Track Monitoring System is
concerned with the inter-state road permit activity i.e. in order to
permit a vehicle to transport across the states. In order to get a
permission to move across the states the owner of the respective
vehicles has to get registered with their respective R.T.A’S to get
permission to move across.
In order to maintain an interaction between different state R.T.A’S.The
main role played here is by the CHECKMASTER who maintain a centralized
database. The details of the vehicle’s and their respective owner
detail’s are stored at their state R.T.A’S database whenever they
register the vehicle by using a local database their process get’s
updated in the centralized once.